refactored spells.ts composable, move constitents to more appropriate files
