+++ app/constants/index.ts
@@ -1,1 +1,2 @@
 export * from './monsters';
+export * from './spells';
