+++ app/helpers/index.ts
@@ -1,4 +1,5 @@
 export { formatModifier } from './formatModifier';
+export { formatSpellSubtitle } from './formatSpellSubtitle';
 export { sortDocumentsByPublisher } from './sortDocumentsByPublisher';
 export { parseChallengeRating } from './parseChallengeRating';
 export { titleCaseToKebabCase } from './titleCaseToKebabCase';
